Increase constructor priority
When running from a C++ environment, constructors of global data under GCC will execute before unthread has been initialized. By increasing the constructor priority, this can be avoided. Before this, I also tried using `.preinit_array` but this ended up not working as `environ` has not been initialized at that time in glibc. Setting the priority to 1000 seems like a hack, but values 0-100 are reserved for the implementation.