Fix memory leaks due to incorrectly set stack ownership
Stacks are only free'd in terminate() if thread->owns_stack is
true. If we malloc'ed a stack, then we own it and should free it.

This commit inverts the meaning of thread->owns_stack so that this
is the case.

Although basic testing shows this fixes the memory leaks, it does
break any attempt to use a pthread_t after it has terminated,
which causes posix tests to break.

Further changes are required.